Split icon.rs
