+++ tower-defense/src/icon/mod.rs
@@ -1,3 +1,6 @@
+mod component;
+mod rendering_tree;
+
 use crate::asset_loader::icon_asset_loader::{IconAssetKind, IconAssetLoader};
 use crate::card::Suit;
 use namui::*;
@@ -165,130 +168,4 @@ pub struct Icon {
 
 fn get_icon_image(_ctx: &RenderCtx, kind: impl Into<IconAssetKind> + Copy) -> Option<namui::Image> {
     IconAssetLoader::get_global().and_then(|loader| loader.get(kind))
-}
-
-impl Component for Icon {
-    fn render(self, ctx: &RenderCtx) {
-        let Self {
-            kind,
-            size,
-            attributes,
-            wh,
-            opacity,
-        } = self;
-        let icon_size = size.px();
-        let icon_wh = Wh {
-            width: icon_size,
-            height: icon_size,
-        };
-        let icon_xy = Xy::new(
-            (wh.width - icon_wh.width) / 2.0,
-            (wh.height - icon_wh.height) / 2.0,
-        );
-        let rect = Rect::from_xy_wh(icon_xy, icon_wh);
-        let image = get_icon_image(ctx, kind);
-        let Some(image) = image else {
-            return;
-        };
-
-        // Create paint with opacity
-        let paint = if opacity < 1.0 {
-            Some(Paint::new(Color::from_f01(1.0, 1.0, 1.0, opacity)))
-        } else {
-            None
-        };
-
-        for attribute in attributes {
-            let attribute_image = get_icon_image(ctx, attribute.icon_kind);
-            let Some(attribute_image) = attribute_image else {
-                continue;
-            };
-            let attribute_render_rect = attribute.attribute_render_rect(rect);
-            ctx.add(namui::image(ImageParam {
-                rect: attribute_render_rect,
-                image: attribute_image.clone(),
-                style: ImageStyle {
-                    fit: ImageFit::Contain,
-                    paint: paint.clone(),
-                },
-            }));
-        }
-        ctx.add(namui::image(ImageParam {
-            rect,
-            image: image.clone(),
-            style: ImageStyle {
-                fit: ImageFit::Contain,
-                paint,
-            },
-        }));
-    }
-}
-
-impl Icon {
-    pub fn to_rendering_tree(&self) -> RenderingTree {
-        let Self {
-            kind,
-            size,
-            attributes,
-            wh,
-            opacity,
-        } = self;
-        let icon_size = size.px();
-        let icon_wh = Wh {
-            width: icon_size,
-            height: icon_size,
-        };
-        let icon_xy = Xy::new(
-            (wh.width - icon_wh.width) / 2.0,
-            (wh.height - icon_wh.height) / 2.0,
-        );
-        let rect = Rect::from_xy_wh(icon_xy, icon_wh);
-
-        let mut rendering_trees = Vec::new();
-
-        // Try to get images from global asset loader
-        if let Some(global_loader) = IconAssetLoader::get_global() {
-            // Add attribute images
-            for attribute in attributes {
-                if let Some(attribute_image) = global_loader.get(attribute.icon_kind) {
-                    let attribute_render_rect = attribute.attribute_render_rect(rect);
-                    let paint = if *opacity < 1.0 {
-                        Some(Paint::new(Color::from_f01(1.0, 1.0, 1.0, *opacity)))
-                    } else {
-                        None
-                    };
-                    rendering_trees.push(namui::image(ImageParam {
-                        rect: attribute_render_rect,
-                        image: attribute_image,
-                        style: ImageStyle {
-                            fit: ImageFit::Contain,
-                            paint: paint.clone(),
-                        },
-                    }));
-                }
-            }
-
-            // Add main icon image
-            if let Some(image) = global_loader.get(*kind) {
-                let paint = if *opacity < 1.0 {
-                    Some(Paint::new(Color::from_f01(1.0, 1.0, 1.0, *opacity)))
-                } else {
-                    None
-                };
-                rendering_trees.push(namui::image(ImageParam {
-                    rect,
-                    image,
-                    style: ImageStyle {
-                        fit: ImageFit::Contain,
-                        paint,
-                    },
-                }));
-            }
-
-            return namui::render(rendering_trees);
-        }
-
-        // Fallback to empty rendering tree if global loader is not available
-        namui::render(rendering_trees)
-    }
 }
